prettify tests
