+++ tests/integration/test_cluster_copier/test.py
@@ -2,13 +2,13 @@ import os
 import random
 import sys
 import time
-from contextlib import contextmanager
-
-import docker
 import kazoo
 import pytest
+from contextlib import contextmanager
 from helpers.cluster import ClickHouseCluster
 from helpers.test_tools import TSV
+
+import docker
 
 CURRENT_TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0, os.path.dirname(CURRENT_TEST_DIR))
